pyproject.toml
[project]
name = "my-mcp-server"
version = "0.1.0"
description = "MCP server for Spotify - top tracks, artists, recently played"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.24.0",
    "spotipy>=2.24.0",
    "python-dotenv>=1.0.0",
]
